Add files via upload
This GUI application facilitates the forecasting of a time series (𝑦) using a Nonlinear Autoregressive model with Exogenous Inputs (NARX), implemented via a feedforward neural network. Users can upload their data as a .mat file, which should include a time series vector (𝑦) and a corresponding matrix of exogenous regressors (𝑋), both having the same number of observations. The interface allows for the customization of key model parameters, including the lag order and the number of hidden layers, enabling flexibility and adaptability to various datasets.